requirements.txt
APScheduler==3.6.3
asgiref==3.5.2
atomicwrites==1.4.1
attrs==22.1.0
cachetools==4.2.2
certifi==2023.5.7
charset-normalizer==2.0.12
colorama==0.4.6
Django==3.2.16
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
et-xmlfile==1.1.0
gpt4all==0.2.3
idna==3.4
iniconfig==1.1.1
isort==5.12.0
lxml==4.9.3
mccabe==0.7.0
openpyxl==3.1.2
packaging==21.3
Pillow==9.3.0
pluggy==0.13.1
py==1.11.0
pycodestyle==2.10.0
pydocstyle==6.3.0
PyJWT==2.1.0
pyparsing==3.0.9
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytest-timeout==2.1.0
python-docx==0.8.11
python-dotenv==0.19.0
python-telegram-bot==13.7
pytz==2022.6
requests==2.26.0
six==1.16.0
snowballstemmer==2.2.0
sqlparse==0.4.3
toml==0.10.2
tornado==6.3.2
tqdm==4.65.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.16
